Add CardServices component with modal and animations
Implemented the CardServices component with a modal feature that opens on card click, using framer-motion for smooth animations. Each card represents a service and includes a 'Leer más' button for further details. Added the CloseIcon for closing the modal and integrated the useOutsideClick hook to close it when clicking outside.